# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Café Lauren",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/recipes": {
      "get": {
        "tags": [
          "recipes"
        ],
        "summary": "List Recipes",
        "operationId": "listRecipes",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search title, description and ingredients",
              "title": "Q"
            },
            "description": "Search title, description and ingredients"
          },
          {
            "name": "filter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Chip: \"All\", \"5 stars\", or a tag such as \"Quick\"",
              "title": "Filter"
            },
            "description": "Chip: \"All\", \"5 stars\", or a tag such as \"Quick\""
          },
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "default": [],
              "title": "Tag"
            }
          },
          {
            "name": "method",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Method"
            }
          },
          {
            "name": "min_stars",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 5,
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Stars"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "saved",
                "draft",
                "all"
              ],
              "type": "string",
              "default": "saved",
              "title": "Status"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "title",
                "stars",
                "last_made",
                "recent"
              ],
              "type": "string",
              "default": "stars",
              "title": "Sort"
            }
          },
          {
            "name": "x-cafe-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "lauren | joe | leidy",
              "title": "X-Cafe-User"
            },
            "description": "lauren | joe | leidy"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecipeSummary"
                  },
                  "title": "Response Listrecipes"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "recipes"
        ],
        "summary": "Create Recipe",
        "operationId": "createRecipe",
        "parameters": [
          {
            "name": "x-cafe-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "lauren | joe | leidy",
              "title": "X-Cafe-User"
            },
            "description": "lauren | joe | leidy"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/recipes/draft": {
      "post": {
        "tags": [
          "recipes"
        ],
        "summary": "Draft Recipe",
        "operationId": "draftRecipe",
        "parameters": [
          {
            "name": "x-cafe-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "lauren | joe | leidy",
              "title": "X-Cafe-User"
            },
            "description": "lauren | joe | leidy"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeDraftRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/recipes/{recipe_id}": {
      "get": {
        "tags": [
          "recipes"
        ],
        "summary": "Get Recipe",
        "operationId": "getRecipe",
        "parameters": [
          {
            "name": "recipe_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Recipe Id"
            }
          },
          {
            "name": "x-cafe-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "lauren | joe | leidy",
              "title": "X-Cafe-User"
            },
            "description": "lauren | joe | leidy"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "recipes"
        ],
        "summary": "Patch Recipe",
        "operationId": "patchRecipe",
        "parameters": [
          {
            "name": "recipe_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Recipe Id"
            }
          },
          {
            "name": "x-cafe-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "lauren | joe | leidy",
              "title": "X-Cafe-User"
            },
            "description": "lauren | joe | leidy"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipePatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "recipes"
        ],
        "summary": "Delete Recipe",
        "operationId": "deleteRecipe",
        "parameters": [
          {
            "name": "recipe_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Recipe Id"
            }
          },
          {
            "name": "x-cafe-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "lauren | joe | leidy",
              "title": "X-Cafe-User"
            },
            "description": "lauren | joe | leidy"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/recipes/{recipe_id}/save": {
      "post": {
        "tags": [
          "recipes"
        ],
        "summary": "Save Recipe",
        "operationId": "saveRecipe",
        "parameters": [
          {
            "name": "recipe_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Recipe Id"
            }
          },
          {
            "name": "x-cafe-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "lauren | joe | leidy",
              "title": "X-Cafe-User"
            },
            "description": "lauren | joe | leidy"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/recipes/{recipe_id}/cooked": {
      "post": {
        "tags": [
          "recipes"
        ],
        "summary": "Cooked",
        "operationId": "markCooked",
        "parameters": [
          {
            "name": "recipe_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Recipe Id"
            }
          },
          {
            "name": "x-cafe-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "lauren | joe | leidy",
              "title": "X-Cafe-User"
            },
            "description": "lauren | joe | leidy"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CookedRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CookedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CookedResponse": {
        "properties": {
          "recipe": {
            "$ref": "#/components/schemas/Recipe"
          },
          "log": {
            "$ref": "#/components/schemas/CookLogEntry"
          }
        },
        "type": "object",
        "required": [
          "recipe",
          "log"
        ],
        "title": "CookedResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobAccepted": {
        "properties": {
          "job": {
            "$ref": "#/components/schemas/Job"
          }
        },
        "type": "object",
        "required": [
          "job"
        ],
        "title": "JobAccepted"
      },
...
      "Ok": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          }
        },
        "type": "object",
        "title": "Ok"
      },
...
      "RecipeCreate": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "short_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Short Title"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "method": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Method"
          },
          "prep_min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prep Min"
          },
          "cook_min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cook Min"
          },
          "total_min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Min"
          },
          "cost_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Usd"
          },
          "healthy": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 10.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Healthy"
          },
          "delicious": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 10.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Delicious"
          },
          "stars": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 5.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Stars"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "ingredients": {
            "items": {
              "$ref": "#/components/schemas/Ingredient"
            },
            "type": "array",
            "title": "Ingredients"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/StepGroup"
            },
            "type": "array",
            "title": "Steps"
          },
          "leftovers": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Leftovers"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "RecipeCreate",
        "description": "Manual recipe entry (no AI). AI drafts go through POST /api/recipes/draft."
      },
      "RecipeDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "short_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Short Title"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "method": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Method"
          },
          "total_min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Min"
          },
          "cost_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Usd"
          },
          "healthy": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Healthy"
          },
          "delicious": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delicious"
          },
          "stars": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stars"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "source": {
            "type": "string",
            "enum": [
              "imported",
              "ai",
              "manual"
            ],
            "title": "Source"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "saved"
            ],
            "title": "Status"
          },
          "last_made": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Made"
          },
          "prep_min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prep Min"
          },
          "cook_min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cook Min"
          },
          "ingredients": {
            "items": {
              "$ref": "#/components/schemas/Ingredient"
            },
            "type": "array",
            "title": "Ingredients"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/StepGroup"
            },
            "type": "array",
            "title": "Steps"
          },
          "leftovers": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Leftovers"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "on_week_days": {
            "items": {
              "type": "string",
              "enum": [
                "mon",
                "tue",
                "wed",
                "thu",
                "fri",
                "sat",
                "sun"
              ]
            },
            "type": "array",
            "title": "On Week Days",
            "description": "Days of the current week using it."
          },
          "in_queue": {
            "type": "boolean",
            "title": "In Queue",
            "default": false
          },
          "ratings": {
            "items": {
              "$ref": "#/components/schemas/CookLogEntry"
            },
            "type": "array",
            "title": "Ratings"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "source",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "RecipeDetail"
      },
      "RecipeDraftRequest": {
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "describe",
              "link",
              "photo",
              "paste"
            ],
            "title": "Mode"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "photo_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Path",
            "description": "Media-relative path of an uploaded photo."
          }
        },
        "type": "object",
        "required": [
          "mode"
        ],
        "title": "RecipeDraftRequest"
      },
...
      "RecipeSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "short_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Short Title"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "method": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Method"
          },
          "total_min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Min"
          },
          "cost_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Usd"
          },
          "healthy": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Healthy"
          },
          "delicious": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delicious"
          },
          "stars": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stars"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "source": {
            "type": "string",
            "enum": [
              "imported",
              "ai",
              "manual"
            ],
            "title": "Source"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "saved"
            ],
            "title": "Status"
          },
          "last_made": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Made"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "source",
          "status"
        ],
        "title": "RecipeSummary"
      },
...
      "StepGroup": {
        "properties": {
          "group": {
            "type": "string",
            "title": "Group"
          },
          "steps": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Steps",
            "description": "Markdown; quantities in **bold**."
          }
        },
        "type": "object",
        "required": [
          "group"
        ],
        "title": "StepGroup"
      },
...
    }
  }
}